Move notification channel creation into main activity. Add notification icon.
