+++ python-for-android/dists/kolibri/src/main/java/org/kivy/android/PythonActivity.java
@@ -1,5 +1,8 @@
 package org.kivy.android;
 
+import android.app.NotificationChannel;
+import android.app.NotificationManager;
+import android.os.Build;
 import android.os.SystemClock;
 
 import java.io.InputStream;
@@ -38,6 +41,9 @@ import android.webkit.WebView;
 import android.webkit.CookieManager;
 import android.net.Uri;
 
+import androidx.core.app.NotificationManagerCompat;
+
+import org.learningequality.Kolibri.R;
 import org.renpy.android.ResourceManager;
 
 public class PythonActivity extends Activity {
@@ -98,6 +104,7 @@ public class PythonActivity extends Activity {
 
         this.mActivity = this;
         this.showLoadingScreen();
+        this.createNotificationChannel();
         new UnpackFilesTask().execute(getAppRoot());
     }
 
@@ -553,6 +560,22 @@ public class PythonActivity extends Activity {
 
     public void requestPermissions(String[] permissions) {
         requestPermissionsWithRequestCode(permissions, 1);
+    }
+
+    private void createNotificationChannel() {
+        // Create the NotificationChannel, but only on API 26+ because
+        // the NotificationChannel class is not in the Support Library.
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            Context context = getApplicationContext();
+            CharSequence name = context.getString(R.string.notification_channel_title);
+            String channelId = context.getString(R.string.notification_channel_id);
+            int importance = NotificationManager.IMPORTANCE_DEFAULT;
+            NotificationChannel channel = new NotificationChannel(channelId, name, importance);
+            // Register the channel with the system. You can't change the importance
+            // or other notification behaviors after this.
+            NotificationManagerCompat notificationManager = NotificationManagerCompat.from(context);
+            notificationManager.createNotificationChannel(channel);
+        }
     }
 }
